Refactor the folder and import structure
