+++ examples/flux_vector/plot_flux_vector_pmsyrm_5kw.py
@@ -15,14 +15,15 @@ does.
 
 from os import path
 import inspect
+
 import numpy as np
 from scipy.io import loadmat
 from scipy.optimize import minimize_scalar
 import matplotlib.pyplot as plt
+
 from motulator.drive import model
 import motulator.drive.control.sm as control
-from motulator.drive.utils import BaseValues, NominalValues, plot
-from motulator.utils import Sequence
+from motulator.drive.utils import BaseValues, NominalValues, plot, Sequence
 
 # %%
 # Compute base values based on the nominal values (just for figures).
