+++ examples/flux_vector/plot_flux_vector_syrm_7kw.py
@@ -14,10 +14,10 @@ Naturally, the PM-flux estimation can be used in PM machine drives as well.
 
 import numpy as np
 import matplotlib.pyplot as plt
+
 from motulator.drive import model
 import motulator.drive.control.sm as control
-from motulator.drive.utils import BaseValues, NominalValues, plot
-from motulator.utils import Sequence
+from motulator.drive.utils import BaseValues, NominalValues, plot, Sequence
 
 # %%
 # Compute base values based on the nominal values (just for figures).
