+++ examples/obs_vhz/plot_obs_vhz_ctrl_pmsyrm_thor.py
@@ -20,14 +20,14 @@ saturation into account in the control algorithm.
 
 from os import path
 import inspect
+
 import numpy as np
 from scipy.optimize import minimize_scalar
 from scipy.interpolate import LinearNDInterpolator
 
 from motulator.drive import model
 import motulator.drive.control.sm as control
-from motulator.drive.utils import BaseValues, NominalValues, plot
-from motulator.utils import Sequence
+from motulator.drive.utils import BaseValues, NominalValues, plot, Sequence
 from motulator.drive.utils import (
     import_syre_data, plot_flux_vs_current, plot_flux_map)
