+++ examples/signal_inj/plot_signal_inj_pmsm_2kw.py
@@ -10,10 +10,10 @@ Square-wave signal injection is used with a simple phase-locked loop.
 
 import numpy as np
 import matplotlib.pyplot as plt
+
 from motulator.drive import model
 import motulator.drive.control.sm as control
-from motulator.drive.utils import BaseValues, NominalValues, plot
-from motulator.utils import Sequence
+from motulator.drive.utils import BaseValues, NominalValues, plot, Sequence
 
 # %%
 # Compute base values based on the nominal values (just for figures).
