+++ examples/vector/plot_vector_ctrl_pmsm_2kw.py
@@ -9,6 +9,7 @@ drive.
 # %%
 
 import time
+
 from motulator.drive import model
 import motulator.drive.control.sm as control
 from motulator.drive.utils import BaseValues, NominalValues, plot
