+++ examples/vector/plot_vector_ctrl_pmsyrm_thor.py
@@ -9,6 +9,7 @@ magnet synchronous reluctance motor. Control look-up tables are also plotted.
 # %%
 
 import numpy as np
+
 from motulator.drive import model
 import motulator.drive.control.sm as control
 from motulator.drive.utils import BaseValues, NominalValues, plot
