--- motulator/_common/_model/__init__.py
@@ -1,12 +1,0 @@
-"""Common functions and classes for continuous-time system models."""
-from motulator._common._model._simulation import (
-    CarrierComparison, Delay, Model, Simulation, Subsystem, zoh)
-
-__all__ = [
-    "CarrierComparison",
-    "Delay",
-    "Model",
-    "Simulation",
-    "Subsystem",
-    "zoh",
-]
