+++ motulator/common/control/_control.py
@@ -5,7 +5,7 @@ from types import SimpleNamespace
 
 import numpy as np
 
-from motulator.utils import abc2complex, complex2abc
+from motulator.common.utils import abc2complex, complex2abc
 
 
 # %%
