+++ motulator/common/utils/__init__.py
@@ -1,5 +1,5 @@
-"""Utility functions."""
-from motulator.utils._helpers import (
+"""Common utilities."""
+from motulator.common.utils._utils import (
     abc2complex, complex2abc, Sequence, Step, wrap)
 
 __all__ = [
