+++ motulator/drive/__init__.py
@@ -1,4 +1,1 @@
 """Drive models and controls."""
-from motulator.drive import control, model, utils
-
-__all__ = ["control", "model", "utils"]
