+++ motulator/drive/control/__init__.py
@@ -1,5 +1,4 @@
 """Controllers for machine drives."""
-from motulator.drive.control import im, sm
 from motulator.drive.control._common import DriveCtrl, SpeedCtrl
 
-__all__ = ["im", "sm", "DriveCtrl", "SpeedCtrl"]
+__all__ = ["DriveCtrl", "SpeedCtrl"]
