+++ motulator/drive/control/im/__init__.py
@@ -1,27 +1,26 @@
 """Controls for induction machines."""
-from motulator.drive.control.im._common import ModelPars, Observer, ObserverCfg
 from motulator.drive.control.im._common import (
-    FullOrderObserver, FullOrderObserverCfg)
+    FullOrderObserver, FullOrderObserverCfg, ModelPars, Observer, ObserverCfg)
 from motulator.drive.control.im._current_vector import (
     CurrentCtrl, CurrentReference, CurrentReferenceCfg, CurrentVectorCtrl)
-from motulator.drive.control.im._vhz import VHzCtrl, VHzCtrlCfg
 from motulator.drive.control.im._obs_vhz import (
     ObserverBasedVHzCtrl, ObserverBasedVHzCtrlCfg)
+from motulator.drive.control.im._vhz import VHzCtrl, VHzCtrlCfg
 from motulator.drive.control._common import SpeedCtrl
 
 __all__ = [
+    "FullOrderObserver",
+    "FullOrderObserverCfg",
     "ModelPars",
     "Observer",
     "ObserverCfg",
-    "FullOrderObserver",
-    "FullOrderObserverCfg",
     "CurrentCtrl",
     "CurrentReference",
     "CurrentReferenceCfg",
     "CurrentVectorCtrl",
-    "VHzCtrl",
-    "VHzCtrlCfg",
     "ObserverBasedVHzCtrl",
     "ObserverBasedVHzCtrlCfg",
+    "VHzCtrl",
+    "VHzCtrlCfg",
     "SpeedCtrl",
 ]
