+++ motulator/drive/control/sm/__init__.py
@@ -1,5 +1,4 @@
 """Controls for synchronous machines."""
-
 from motulator.drive.control.sm._common import ModelPars, Observer, ObserverCfg
 from motulator.drive.control.sm._flux_vector import (
     FluxTorqueReference, FluxTorqueReferenceCfg, FluxVectorCtrl)
@@ -9,8 +8,8 @@ from motulator.drive.control.sm._obs_vhz import (
     ObserverBasedVHzCtrl, ObserverBasedVHzCtrlCfg)
 from motulator.drive.control.sm._signal_inj import (
     SignalInjection, SignalInjectionCtrl)
-from motulator.drive.control._common import SpeedCtrl
 from motulator.drive.control.sm._torque import TorqueCharacteristics
+from motulator.drive.control._common import SpeedCtrl
 
 __all__ = [
     "ModelPars",
@@ -27,6 +26,6 @@ __all__ = [
     "ObserverBasedVHzCtrlCfg",
     "SignalInjection",
     "SignalInjectionCtrl",
-    "SpeedCtrl",
     "TorqueCharacteristics",
+    "SpeedCtrl",
 ]
