+++ motulator/drive/control/sm/_common.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass, field, InitVar
 
 import numpy as np
 
-from motulator.utils import wrap
+from motulator.common.utils import wrap
 
 
 # %%
