+++ motulator/drive/control/sm/_current_vector.py
@@ -4,8 +4,8 @@ from dataclasses import dataclass, InitVar
 
 import numpy as np
 
-from motulator.drive.control._common import DriveCtrl, SpeedCtrl
-from motulator._common._control import ComplexPICtrl
+from motulator.drive.control import DriveCtrl, SpeedCtrl
+from motulator.common.control import ComplexPICtrl
 from motulator.drive.control.sm._common import ModelPars, Observer, ObserverCfg
 from motulator.drive.control.sm._torque import TorqueCharacteristics
