+++ motulator/drive/model/__init__.py
@@ -1,8 +1,8 @@
 """Continuous-time machine drive models."""
-from motulator._common._model._simulation import (
+from motulator.common.model._simulation import (
     CarrierComparison, Delay, Simulation)
 from motulator.drive.model._drive import Drive, DriveWithLCFilter
-from motulator._common._model._converter import FrequencyConverter, Inverter
+from motulator.common.model._converter import FrequencyConverter, Inverter
 from motulator.drive.model._machine import (
     InductionMachine, InductionMachineInvGamma, SynchronousMachine)
 from motulator.drive.model._mechanics import Mechanics, TwoMassMechanics
